creation and editing of lists
Lists will be stored as a two dimensional array in a separate file, much like alarms and the calendar.or the creation of lists I have a long series of synonyms for ‘named’ that the program will scan for The word that directly follows the last appearance of one of these synonyms will be the name of a given list.
The program also checks to see if the supplied name is duplicate of an existing name to avoid having multiple entries that require more complex code to distinguish.

Editing the lists works in a similar fashion. The program searches the command for the list title and then adds everything to the right of that as an entry into the list array.